flask_app/requirements.txt
Flask==2.2.3
keras==2.11.0
numpy==1.24.2
Pillow==9.4.0
requests==2.28.2
torch==1.13.1
transformers==4.27.1
tensorflow==2.11.0
